+++ lib/webspicy/client/http_client.rb
@@ -7,7 +7,9 @@ module Webspicy
     end
     attr_reader :api
 
-    def call(test_case, service, resource)
+    def call(test_case)
+      service, resource = test_case.service, test_case.resource
+
       # Instantiate the parameters
       headers = test_case.headers
       params = test_case.dress_params? ? service.dress_params(test_case.params) : test_case.params