+++ lib/webspicy/tester.rb
@@ -12,18 +12,18 @@ module Webspicy
         client = scope.get_client
         scope.each_resource do |resource|
           scope.each_service(resource) do |service|
-            rspec_service!(service, resource, client, scope)
+            rspec_service!(service, client, scope)
           end
         end
       end
       RSpec::Core::Runner.run config.rspec_options
     end
 
-    def rspec_service!(service, resource, client, scope)
-      RSpec.describe "#{service.method} #{resource.url}" do
-        scope.each_testcase(service, resource) do |test_case, counterexample|
+    def rspec_service!(service, client, scope)
+      RSpec.describe service do
+        scope.each_testcase(service) do |test_case, counterexample|
           describe test_case do
-            include_examples 'a successful test case invocation', client, test_case, service, resource, counterexample
+            include_examples 'a successful test case invocation', client, test_case, counterexample
           end
         end
       end
@@ -31,13 +31,13 @@ module Webspicy
 
     def rspec_config!
       return if @rspec_config
-      RSpec.shared_examples "a successful test case invocation" do |client, test_case, service, resource, counterexample|
+      RSpec.shared_examples "a successful test case invocation" do |client, test_case, counterexample|
 
         before(:all) do
           @invocation ||= begin
-            service.instrument(test_case, service, resource)
-            client.before(test_case, service, resource)
-            client.call(test_case, service, resource)
+            test_case.instrument
+            client.before(test_case)
+            client.call(test_case)
           end
         end
 
@@ -59,7 +59,7 @@ module Webspicy
 
         it 'meets declarative postconditions' do
           expect(invocation.postconditions_unmet).to(be_nil)
-        end if service.has_postconditions? and not(counterexample)
+        end if test_case.service.has_postconditions? and not(counterexample)
 
         it 'meets the specific error messages, if any (backward compatibility)' do
           expect(@invocation.expected_error_unmet).to(be_nil)